split statistics test file
